fix: modify header elements for cross-browser compatibility
The yellow headers where spacing differently in different browsers with
the use of the <br /> take. Could not find a solution that could add a
height to the <br /> element that consistently worked across all
browsers so settled with converting each line to a <p> tag and creating
the yellow background with a <span> element.